Refactor: 차단 기능 관련 코드 별도의 클래스로 분리
