+++ healthper/src/main/java/umc/healthper/service/MemberService.java
@@ -4,13 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.healthper.domain.member.Member;
-import umc.healthper.domain.member.MemberBlock;
 import umc.healthper.domain.member.MemberStatus;
 import umc.healthper.dto.member.UpdateMemberRequestDto;
-import umc.healthper.exception.block.MemberBlockDuplicateException;
-import umc.healthper.exception.block.MemberBlockNotFoundException;
 import umc.healthper.exception.member.*;
-import umc.healthper.repository.MemberBlockRepository;
 import umc.healthper.repository.MemberRepository;
 
 import java.util.List;
@@ -22,7 +18,6 @@ import java.util.Optional;
 public class MemberService {
 
     private final MemberRepository memberRepository;
-    private final MemberBlockRepository memberBlockRepository;
 
     /**
      * 회원 등록
@@ -114,47 +109,6 @@ public class MemberService {
     }
 
     /**
-     * 회원 차단
-     *
-     * @param loginMemberId   로그인 Member의 id (주체)
-     * @param blockedMemberId 차단할 Member의 id (대상)
-     * @throws MemberNotFoundByIdException   전달받은 id에 해당하는 Member가 존재하지 않는 경우
-     * @throws MemberBlockDuplicateException 이미 차단한 Member인 경우
-     */
-    public void blockMember(Long loginMemberId, Long blockedMemberId) {
-        Member loginMember = memberRepository.findById(loginMemberId)
-                .orElseThrow(MemberNotFoundByIdException::new);
-        Member blockedMember = memberRepository.findById(blockedMemberId)
-                .orElseThrow(MemberNotFoundByIdException::new);
-
-        validateDuplicateBlockedMember(loginMember, blockedMember);
-
-        MemberBlock memberBlock = MemberBlock.createMemberBlock(loginMember, blockedMember);
-        memberBlockRepository.save(memberBlock);
-    }
-
-    /**
-     * 차단했던 회원의 차단 취소
-     *
-     * @param loginMemberId   로그인 Member (주체)
-     * @param blockedMemberId 차단했던 Member (대상)
-     * @throws MemberNotFoundByIdException  전달받은 id에 해당하는 Member가 존재하지 않는 경우
-     * @throws MemberBlockNotFoundException 차단한 적 없는 Member인 경우
-     */
-    public void deleteBlockedMember(Long loginMemberId, Long blockedMemberId) {
-        Member loginMember = memberRepository.findById(loginMemberId)
-                .orElseThrow(MemberNotFoundByIdException::new);
-        Member blockedMember = memberRepository.findById(blockedMemberId)
-                .orElseThrow(MemberNotFoundByIdException::new);
-
-        MemberBlock memberBlock = memberBlockRepository.findByMemberAndBlockedMember(loginMember, blockedMember)
-                .orElseThrow(MemberBlockNotFoundException::new);
-
-        loginMember.getMemberBlocks().remove(memberBlock);
-        memberBlockRepository.delete(memberBlock);
-    }
-
-    /**
      * 이미 사용중인 닉네임인지 검증
      *
      * @param nickname 사용하고자 하는 닉네임
@@ -175,21 +129,6 @@ public class MemberService {
     private void validAlreadyResignedMember(Member member) {
         if (member.getStatus() == MemberStatus.RESIGNED) {
             throw new AlreadyResignedMemberException("이미 탈퇴한 회원입니다.");
-        }
-    }
-
-    /**
-     * 이미 차단한 회원인지 중복 차단 여부 검증
-     *
-     * @param member        로그인 Member (주체)
-     * @param blockedMember 차단할 Member (대상)
-     * @throws MemberBlockDuplicateException 이미 차단한 Member인 경우
-     */
-    private void validateDuplicateBlockedMember(Member member, Member blockedMember) {
-        for (MemberBlock memberBlock : member.getMemberBlocks()) {
-            if (memberBlock.getBlockedMember() == blockedMember) {
-                throw new MemberBlockDuplicateException("이미 차단한 회원입니다.");
-            }
         }
     }
 }
